requirements.txt
numpy
tqdm
pyserial